[ALVS] state sync: fixes from unit tests
redundant zeroing of ipv4 checksum field.
change current buf len field type to 16bit to prevent len=256 overflow.

+++ src/dp/alvs_defs.h
@@ -90,7 +90,7 @@ struct alvs_conn_sync_state {
 	enum alvs_conn_sync_status        conn_sync_status:4;
 	unsigned                          amount_buffers:4;
 	uint8_t                           *current_base;
-	uint8_t                           current_len;
+	uint16_t                           current_len;
 	uint8_t                           conn_count;
 };
